Add core_line_change_recent_year params (#33)
Whats the core code of project? I think the answer is that if the
project is written by a specific language, then these specific
language code is the core code.

This patch adds a `core_line_change_recent_year` to calculate the
line change of core code. The result like below:

"+661, -174 (*.py: +627, -174 *.cfg: +34, -0)"

means that "There are 661 loc additon and 174 loc deletion, include
627 additon/174 deletion python code, and 34 addtion cfg code".